meta: refactor uefi-signing
- Encapsulate knowledge of local signing by moving most TEGRA_UEFI_DB_*
  references to tegra-uefi-signing.bbclass.

- All recipes just call the tegra_uefi_* signing functions.

- Recipes that require uefi signing inherit ${TEGRA_UEFI_SIGNING_CLASS}
  which defaults to tegra-uefi-signing.bbclass.  Users can assign
  their own class to override the default uefi signing function
  behavior (to use a signing server, as an example).

- Use weak default assignment for TEGRA_UEFI_SIGNING_CLASS and
  TEGRA_UEFI_USE_SIGNED_FILES in the event they are not defined.

+++ classes-recipe/tegra-uefi-signing.bbclass
@@ -10,23 +10,30 @@ TEGRA_UEFI_DB_KEY ??= ""
 TEGRA_UEFI_DB_CERT ??= ""
 TEGRA_UEFI_SIGNING_TASKDEPS ?= "${@tegra_uefi_signing_deps(d, tasks=True)}"
 TEGRA_UEFI_SIGNING_DEPENDS ?= "${@tegra_uefi_signing_deps(d)}"
+TEGRA_UEFI_USE_SIGNED_FILES ?= "${@'true' if d.getVar('TEGRA_UEFI_DB_KEY') and d.getVar('TEGRA_UEFI_DB_CERT') else 'false'}"
 
 # Standard signing, input file modified with signature
 tegra_uefi_sbsign() {
-    sbsign --key "${TEGRA_UEFI_DB_KEY}" --cert "${TEGRA_UEFI_DB_CERT}" --output "$1" "$1"
+    if [ -n "${TEGRA_UEFI_DB_KEY}" -a -n "${TEGRA_UEFI_DB_CERT}" ]; then
+        sbsign --key "${TEGRA_UEFI_DB_KEY}" --cert "${TEGRA_UEFI_DB_CERT}" --output "$1" "$1"
+    fi
 }
 
 # Separate signature file, for NVIDIA's L4TLauncher 
 tegra_uefi_split_sign() {
-    openssl cms -sign -signer "${TEGRA_UEFI_DB_CERT}" -inkey "${TEGRA_UEFI_DB_KEY}" -binary -in "$1" -outform der -out "$1".sig
+    if [ -n "${TEGRA_UEFI_DB_KEY}" -a -n "${TEGRA_UEFI_DB_CERT}" ]; then
+        openssl cms -sign -signer "${TEGRA_UEFI_DB_CERT}" -inkey "${TEGRA_UEFI_DB_KEY}" -binary -in "$1" -outform der -out "$1".sig
+    fi
 }
 
 # Signature attached to end, another NVIDIA special
 # Input file remains intact; output file has ".signed" suffix
 tegra_uefi_attach_sign() {
-    openssl cms -sign -signer "${TEGRA_UEFI_DB_CERT}" -inkey "${TEGRA_UEFI_DB_KEY}" -binary -in "$1" -outform der -out "$1".sig.tmp
     cp "$1" "$1.signed"
-    truncate --size=%2048 "$1.signed"
-    cat "$1".sig.tmp >> "$1.signed"
-    rm "$1".sig.tmp
+    if [ -n "${TEGRA_UEFI_DB_KEY}" -a -n "${TEGRA_UEFI_DB_CERT}" ]; then
+        openssl cms -sign -signer "${TEGRA_UEFI_DB_CERT}" -inkey "${TEGRA_UEFI_DB_KEY}" -binary -in "$1" -outform der -out "$1".sig.tmp
+        truncate --size=%2048 "$1.signed"
+        cat "$1".sig.tmp >> "$1.signed"
+        rm "$1".sig.tmp
+    fi
 }
